Segment transfer flow into SendView -> SendForm -> SendTransferForm hierarchy
